+++ tests/test_custom.py
@@ -20,7 +20,7 @@ def test_custom_item_remote_schema_v080():
             "validation_method": "custom",
             "valid_stac": False,
             "error_type": "JSONSchemaValidationError",
-            "error_message": "'bbox' is a required property of the root of the STAC object",
+            "error_message": "'bbox' is a required property",
         }
     ]
 
@@ -74,7 +74,7 @@ def test_custom_bad_item_remote_schema_v090():
             "schema": ["https://cdn.staclint.com/v0.9.0/item.json"],
             "valid_stac": False,
             "error_type": "JSONSchemaValidationError",
-            "error_message": "'id' is a required property of the root of the STAC object",
+            "error_message": "'id' is a required property",
         }
     ]
 
@@ -95,29 +95,6 @@ def test_custom_item_remote_schema_v1rc2():
                 "https://schemas.stacspec.org/v1.0.0-rc.2/item-spec/json-schema/item.json"
             ],
             "valid_stac": True,
-        }
-    ]
-
-
-def test_custom_proj_error_v1rc2():
-    schema = "https://stac-extensions.github.io/projection/v1.0.0/schema.json"
-    stac_file = (
-        "tests/test_data/1rc2/extensions-collection/./proj-example/proj-example.json"
-    )
-    stac = stac_validator.StacValidate(stac_file, custom=schema)
-    stac.run()
-    assert stac.message == [
-        {
-            "version": "1.0.0-rc.2",
-            "path": "tests/test_data/1rc2/extensions-collection/./proj-example/proj-example.json",
-            "schema": [
-                "https://stac-extensions.github.io/projection/v1.0.0/schema.json"
-            ],
-            "valid_stac": False,
-            "asset_type": "ITEM",
-            "validation_method": "custom",
-            "error_type": "JSONSchemaValidationError",
-            "error_message": "'A' is not of type 'number'. Error is in properties -> proj:centroid -> lat ",
         }
     ]