+++ tests/test_recursion.py
@@ -328,7 +328,7 @@ def test_recursion_with_bad_item():
             ],
             "valid_stac": False,
             "error_type": "JSONSchemaValidationError",
-            "error_message": "'id' is a required property of the root of the STAC object",
+            "error_message": "'id' is a required property",
         },
     ]
 
@@ -350,6 +350,6 @@ def test_recursion_with_missing_collection_link():
             "valid_stac": False,
             "validation_method": "recursive",
             "error_type": "JSONSchemaValidationError",
-            "error_message": "'simple-collection' should not be valid under {}. Error is in collection",
+            "error_message": "'simple-collection' should not be valid under {}",
         },
     ]