+++ CHANGELOG.md
@@ -12,6 +12,10 @@ The format is (loosely) based on [Keep a Changelog](http://keepachangelog.com/) 
 - Configure whether to open URLs when validating assets [#238](https://github.com/stac-utils/stac-validator/pull/238)
 - Allow to provide HTTP headers [#239](https://github.com/stac-utils/stac-validator/pull/239)
 
+### Changed
+
+- Switched to the referencing library for dynamic JSON schema validation and reference resolution [#241](https://github.com/stac-utils/stac-validator/pull/241)
+
 ## [v3.4.0] - 2024-10-08
 
 ### Added